Refactor app directory
